Adds NavBar and Footer components and cleans up App.vue
