+++ src/components/Footer.vue
@@ -1,0 +1,80 @@
+<template>
+  <footer class="footer footer-black footer-big">
+    <div class="container">
+      <div class="row">
+        <div class="col-md-4 col-sm-4 ml-auto mr-auto">
+          <div class="logo text-center">
+            <h3>Sparq Fitness</h3>
+          </div>
+        </div>
+        <div class="col-md-6 offset-md-2 col-sm-8">
+          <div class="links">
+            <ul>
+              <li>
+                <a href="#paper-kit">
+                  Home
+                </a>
+              </li>
+              <li>
+                <a href="#paper-kit">
+                  Company
+                </a>
+              </li>
+              <li>
+                <a href="#paper-kit">
+                  Portfolio
+                </a>
+              </li>
+              <li>
+                <a href="#paper-kit">
+                  Team
+                </a>
+              </li>
+              <li>
+                <a href="#paper-kit">
+                  Contact
+                </a>
+              </li>
+              <li>
+                <a href="#paper-kit">
+                  We're Hiring!
+                </a>
+              </li>
+            </ul>
+            <hr />
+            <div class="copyright">
+              <div class="pull-left">
+                ©2019
+              </div>
+              <div class="pull-right">
+                <ul>
+                  <li>
+                    <a href="#paper-kit">
+                      Terms
+                    </a>
+                  </li>
+                  |
+                  <li>
+                    <a href="#paper-kit">
+                      Privacy
+                    </a>
+                  </li>
+                </ul>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </footer>
+</template>
+
+<script>
+  export default {
+    
+  }
+</script>
+
+<style scoped>
+
+</style>
