+++ src/views/trainers/TrainersShow.vue
@@ -18,12 +18,7 @@
               <div id="carousel" class="ml-auto mr-auto">
                 <div class="card page-carousel">
                   <div id="carouselExampleIndicators" class="carousel slide" data-ride="carousel">
-<!--                     <ol class="carousel-indicators">
-                      <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active"></li>
-                      <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
-                      <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
-                      <li data-target="#carouselExampleIndicators" data-slide-to="3"></li>
-                    </ol> -->
+
                     <div class="carousel-inner" role="listbox">
                       <div class="carousel-item active">
                         <img class="d-block img-fluid" :src="trainer.avatar" alt="Awesome Item">
@@ -48,7 +43,76 @@
               </div>
             </div>
           </div>
+
         </div>
+    <div class="projects-2 section section-dark">
+      <div class="container">
+        <div class="row">
+          <div class="col-md-8 ml-auto mr-auto text-center">
+            <h6 class="text-muted">Our work</h6>
+            <h2 class="title">Some of Our Awesome Products - 2</h2>
+            <h5 class="description">This is the paragraph where you can write more details about your projects. Keep you user engaged by providing meaningful information.</h5>
+          </div>
+        </div>
+        <div class="space-top"></div>
+        <div class="row">
+          <div class="col-md-4">
+            <div class="card card-plain">
+              <div class="card-img-top">
+                <a href="http://www.creative-tim.com/product/paper-kit">
+                  <img class="img" src="/assets/img/sections/opt_pk_thumbnail.jpg" />
+                </a>
+              </div>
+              <div class="card-body">
+                <a href="http://www.creative-tim.com/product/paper-kit" target="_blank">
+                  <h4 class="card-title">Paper Kit Free</h4>
+                </a>
+                <h6 class="card-category text-muted">Free Ui kit</h6>
+                <p class="card-description">
+                  Paper Kit is a free Bootstrap UI kit that can get you going in a new creative direction! To say the least, it's different since it's not material, flat or iOS inspired.
+                </p>
+              </div>
+            </div>
+          </div>
+          <div class="col-md-4">
+            <div class="card card-plain">
+              <div class="card-img-top">
+                <a href="https://www.creative-tim.com/product/light-bootstrap-dashboard-pro">
+                  <img class="img" src="/assets/img/sections/opt_lbd_pro_new_thumbnail.jpg" />
+                </a>
+              </div>
+              <div class="card-body">
+                <a href="https://www.creative-tim.com/product/light-bootstrap-dashboard-pro" target="_blank">
+                  <h4 class="card-title">Light Bootstrap Dashboard </h4>
+                </a>
+                <h6 class="card-category text-muted">Premium template</h6>
+                <p class="card-description">
+                  Light Bootstrap Dashboard PRO is a Bootstrap Admin Theme designed to look simple and beautiful. Forget about boring dashboards and grab yourself a copy to kickstart new project!
+                </p>
+              </div>
+            </div>
+          </div>
+          <div class="col-md-4">
+            <div class="card card-plain">
+              <div class="card-img-top">
+                <a href="https://www.creative-tim.com/product/get-shit-done-pro">
+                  <img class="img" src="/assets/img/sections/opt_gsdk_new_thumbnail.jpg" />
+                </a>
+              </div>
+              <div class="card-body">
+                <a href="https://www.creative-tim.com/product/get-shit-done-pro" target="_blank">
+                  <h4 class="card-title">Get Shit Done Kit Pro</h4>
+                </a>
+                <h6 class="card-category text-muted">Premium UI kit</h6>
+                <p class="card-description">
+                  Get Shit Done Kit Pro it's a Bootstrap Kit that comes with a huge number of customisable components. They are pixel perfect, light and easy to use and combine with other elements.
+                </p>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
       </div>
     </div>
     <beautiful-chat
